API references

NumDisplay.DisplayBCD — Type

mutable struct DisplayBCD <: DisplayNoBuffer
    digits_pins::AbstractVector{Int}      # pin numbers to control digits: [LSD ... MSD]
    sectors_pins::Tuple{Int,Int,Int,Int,Int,Int,Int,Int}  # pin numbers for binary code: (d0,d1,d2,d3,-,-,-,DP)
    buffer::AbstractVector{UInt8}         # internal storage of digits values
    usDelay::Int                          # duration of period when digit is on
    inverted_digits::Bool                 # if digit pins states must be inverted, i.e. 1 means LOW pin state
    inverted_sectors::Bool                # if input pins and dp pin states must be inverted, i.e. 1 means LOW pin state
    intensity::Int                        # 16 max
    limit::Int
    test_mode::Bool
end

NumDisplay.DisplayBCD — Type

function DisplayBCD(
    digits_pins::AbstractVector{Int},
    sectors_pins::Tuple{Int,Int,Int,Int,Int}, # A, B, C, D, DP
    scan_rate::Real = 800,                    # Hz
    inverted_digits::Bool = false,
    inverted_sectors::Bool = false
)

Creates device representing numerical display with several digits under control of the BCD chip. The number of display digits equal to digits_pins count.

Arguments

digits_pins : Vector of GPIO pin numbers connected to anode. The HIGH state means the digit is on. LOW means off. The first pin in array should manage the less significant digit.

sectors_pins : Tuple of length 5 consisting of GPIO numbers representing the 4-bit code of a digit. The sequence of pins is the following: (d0, d1, d2, d3, DP).

scan_rate : refresh rate of digits in Hz. The digits in display are controlled by impulses of digits_pins. This argument sets the time period for display of one digit. If scan_rate=1000 the width will be recalculated as 1/1000 = 1e-3 second or 1e3 microsecond. The default value is 800 Hz.

inverted_digits : This option inverts digit_pins active states.

inverted_sectors : This option inverts sectors_pins active states.

NumDisplay.DisplayDirect — Type

mutable struct DisplayDirect <: DisplayNoBuffer
    digits_pins::AbstractVector{Int}             # pins' numbers to control digit: [LSD ... MSD]
    sectors_pins::Tuple{Int,Int,Int,Int,Int,Int,Int,Int} # pins' number to control sectors: (g,f,e,d,c,b,a,DP) | (d0,d1,d2,d3,-,-,-,DP)
    buffer::AbstractVector{UInt8}                # storage for current digit values
    usDelay::Real                                # duration of one active digit
    inverted_digits::Bool                        # if digit pins states must be inverted, i.e. 1 means LOW pin state
    inverted_sectors::Bool                       # if sectors pins' states must be inverted, i.e. 1 means LOW pin state
    decode_mode::UInt8                           # current decode mode for digits, 0b00000001 - only first digit is in decode mode
    intensity::Int                               # value from 1 to 16, brightness controlled by pulse-width
    limit::Int                                   # current number of digits to show
end

NumDisplay.DisplayDirect — Method

function DisplayDirect(
    digits_pins::AbstractVector{Int},
    sectors_pins::Tuple{Int,Int,Int,Int,Int,Int,Int,Int};
    scan_rate::Real = 800, # Hz
    inverted_digits::Bool = false,
    inverted_sectors::Bool = false
)

Creates numerical display device controlled directly with the RaspberryPi pins. The number of display digits equal to digits_pins count. DisplayDirect stores its states in internal buffer and creates the pulse-width wave to show digits on a physical device.

Initial state of display:

shutdown mode on. Use method shutdown_mode_off() to activate the display before the first use.

test mode off.

decode mode is on for all digits.

limit value is equal to all available digits, see size()

intensity = 5

buffer is zero for all digits.

Arguments

digits_pins : Vector of GPIO pin numbers connected to common anode or cathode. The first pin in array manages the less significant digit (LSD) of display. The value -1 is also possible here which means that the digit will not be used.

sectors_pins : Tuple of length 8 consisting of GPIO numbers controlling the states of 8 sectors. The sequence of pins is the following: (g, f, e, d, c, b, a, DP). This corresponds to the sequence of bits (starting from most significant) in buffer. The value -1 is also possible here which means that the pin will not be used.

scan_rate : refresh rate of digits in Hz. The digits in display are controlled by impulses of digits_pins. This argument sets the time period for display of one digit. If scan_rate=1000 the width will be recalculated as 1/1000 = 1e-3 second or 1e3 microsecond. The default value is 800 Hz.

inverted_digits : This option inverts digit_pins active states.

inverted_sectors : This option inverts sectors_pins active states.

NumDisplay.DisplaySPI — Type

struct DisplaySPI <: AbstractNumDisplay
    handle::Int
end

NumDisplay.DisplaySPI — Method

function DisplaySPI(;
    spi::Int = 0,
    channel::Int = 0,
    boud::Int = 32*10^3
)

Creates device that refers to numerical display integrated with SPI chip.

Arguments

spi : value 0 or 1 to use main or auxilary RaspberryPi's SPI controller. Default is main.

channel : number of CE port: 0, 1 for the main SPI or 0, 1, 2 for the auxilary one. Default is 0.

boud : data transfer rate for SPI, bits per second. Must be from 32kHz to 10MHz.

NumDisplay.decode_mode — Function

function decode_mode(d::AbstractNumDisplay, decode::UInt8 = 0b1111_1111)

The decode-mode register sets BCD code (0-9, E, H, L, P, and -) or no-decode operation for each digit.

When the code B decode mode is used, the decoder looks only at the lower nibble of the data in the digit registers (D4–D1), disregarding bits D5–D7. D8 sets the decimal point (DP).

When no-decode is selected, data bits D8–D1 correspond to the segment lines

all digits in decode mode => 0xff = 0b11111111 all digits in none-decode mode => 0x00 = 0b00000000 first decode other none mode => 0x01 = 0b0000_0001

Arguments

d : device instance

decode : Each bit in the register corresponds to one digit. A logic high selects code decoding while logic low bypasses the decoder.

NumDisplay.scan_rate — Method

scan_rate(d::AbstractNumDisplay)

Getter to display actual refresh rate.

Return

Int value in Hz

NumDisplay.set_intensity — Function

function set_intensity(d::AbstractNumDisplay, intensity::Int = 16)

Digital control of display brightness is provided by an internal pulse-width modulator.

Arguments

d : device instance

intensity : the value from 1 to 16 controlling the brighness. The modulator scales the average segment current in 16 steps.

NumDisplay.set_limit — Function

function set_limit(d::AbstractNumDisplay, limit::Int = size(d))

The function sets how many digits are displayed. Since the number of scanned digits affects the display brightness, the scan-limit register should not be used to blank portions of the display.

The default value equal to the maximal value.

Arguments

d : device instance

limit : number of active digits, from 1 to size(d) 1 means only digit 1 is shown. 4 meand 8 digits will be shown.

NumDisplay.shutdown_mode_off — Method

shutdown_mode_off(d::AbstractNumDisplay)

Sets the normal mode, i.e. display digits. It should be run after initialization of device.

NumDisplay.shutdown_mode_on — Method

shutdown_mode_on(d::AbstractNumDisplay)

In shutdown mode you can update digits values and set different settings but nothing will be shown. Shutdown can be used to save power or as an alarm to flash the display by successively entering and leaving shutdown mode.

NumDisplay.size — Method

size(d::AbstractNumDisplay)

Getter to show the maximal number of digits available for the display. This is equal to the number of digits_pins.

NumDisplay.test_mode_off — Method

function test_mode_off(d::AbstractNumDisplay)

Sets the normal mode, i.e. display digits.

NumDisplay.test_mode_on — Method

function test_mode_on(d::AbstractNumDisplay)

Display-test mode turns all LEDs on. Test mode overrides shutdown mode. Display remain in display-test mode until the display is reconfigured for normal operation.

NumDisplay.update — Method

function update(d::DisplayNoBuffer)

Generates a sequence of pulses based on buffer, decode_mode, intensity, limit, and runs pulses repeatedly. The function is used internally after buffer and modes changes. If a display is in shutdown mode, the function does nothing.

NumDisplay.write_digit — Method

write_digit(
    d::AbstractNumDisplay,
    value::UInt8,
    position::Int
)

Writes a digit to the position. The result of the execution is changing one digit.

The result will depend on decode mode, see decode_mode() function. The most significant bit sets the DP (dot) state.

In non-decode mode the bits 7-0 manage the sectors a, b, c, d, e, f, g.

In decode mode only bits 3-0 play a role. The sequence of numbers is the following: 0,1,2,3,4,5,6,7,8,9,-,E,H,L,P,blank

Arguments

d : object representing display device

value : UInt8 value updating the buffer value.

position : number of digit to write starting from 1 which mean least significant digit. The maximal value depends on available digits, so it should be <= length(indicator.digit_pins)

NumDisplay.write_digit — Method

function write_digit(
    indicator::DisplayBCD,
    digit::Union{UInt8, Nothing},
    position::Int
)

Writes a digit to the position. The result of the execution is changing one digit in a particular digit.

Arguments

indicator : object representing display device

digit : decimal value from 0 to 9 or nothing. The last means an empty digit. Values from 10 to 14 are also possible here but results to miningless symbols. Value 15 means an empty digit and it is the same as nothing.

position : number of digit to write starting from 1 which mean less significant digit. The maximal value depends on available digits, so it should be <= length(indicator.digit_pins)

NumDisplay.write_number — Function

write_number(
    indicator::AbstractNumDisplay,
    number::Int,
    dp_position::Union{Int, Nothing} = nothing
)

Writes the decimal value to the display.

Arguments

indicator : object representing display device

number : decimal number. The maximal possible value depends on number of digits and will be checked.

dp_position : position of digit starting from less significant digit or nothing.

Example

# d is 4-digit display
write_number(d, 123)
# the result is _123

write_number(d, 1234)
# the result is 1234

write_number(d, 12345)
# throws an error

NumDisplay.write_number — Method

Writes several digits Runs display in decode mode and display several digits element in numer_vector must be Integers

NumDisplay.write_symbols — Method

works in segment-wise mode set active segments DP A B C D E F G