API references

NumDisplay.DisplayBCD — Type

struct DisplayBCD <: AbstractNumDisplay
    sectors_pins::AbstractVector{Int}   # pin numbers starting from less significant
    input_pins::Tuple{Int,Int,Int,Int}  # pin numbers for binary code starting from less significant
    dp_pin::Union{Int, Nothing}         # pin changing dot state
    buffer::AbstractVector{UInt8}       # internal storage of digits values
    dp_buffer::AbstractVector{UInt8}
    usDelay::Int                        # duration of period when sector is on
end

NumDisplay.DisplayBCD — Type

function DisplayBCD(
    sectors_pins::AbstractVector{Int},
    input_pins::Tuple{Int,Int,Int,Int},
    dp_pin::Union{Int, Nothing} = nothing;
    fps::Int = 1000,
    inverted_sectors::Bool = false,
    inverted_input::Bool = false
)

Creates device representing numerical display with several digits under control of the BCD chip. The number of display digits equal to sectors_pins count.

Arguments

sectors_pins : Vector of GPIO pin numbers connected to anode. The HIGH state means the digit is on. LOW means off. The first pin in array should manage the less significant digit.

input_pins : Tuple consisting of GPIO numbers representing the 4-bit code of a digit. The first pin in tuple is less significant number.

dp_pin : Number of pin connected to dot LED (DP)

fps : frame rate (frames per second). The digits in display are controlled by impulses of sectors_pins. This argument sets the width of one impuls. If fps=1000 the width will be recalculated as 1/1000 = 1e-3 second or 1e3 microsecond. The default value is 1000 Hz.

NumDisplay.clean — Method

clean(indicator::AbstractNumDisplay)

Display empty sectors.

Arguments

indicator : object representing display device

NumDisplay.stop — Method

stop(indicator::AbstractNumDisplay)

Stop device and reset the the initial state.

Arguments

indicator : object representing display device

NumDisplay.update — Method

update(indicator::AbstractNumDisplay)

Show the content of indicator.buffer to the display. This function is used internaly by write... methods to update the display.

Arguments

indicator : object representing display device

NumDisplay.write_digit — Method

function write_digit(
    indicator::DisplayBCD,
    digit::Union{UInt8, Nothing},
    position::Int
)

Writes a digit to the position. The result of the execution is changing one digit in a particular sector.

Arguments

indicator : object representing display device

digit : decimal value from 0 to 9 or nothing. The last means an empty sector. Values from 10 to 14 are also possible here but results to miningless symbols. Value 15 means an empty sector and it is the same as nothing.

position : number of sector to write starting from 1 which mean less signifacant digit. The maximal value depends on available sectors, so it should be <= length(indicator.digit_pins)

NumDisplay.write_dp — Method

function write_dp(
    indicator::AbstractNumDisplay,
    dp_value::UInt8,
    dp_position::Int
)

Writes a dot symbol to the position. The result of the execution is changing one dot in a particular sector.

Arguments

indicator : object representing display device

dp_value : value 0 or 1 to get dot OFF or ON.

position : number of sector to write starting from 1 which mean less signifacant digit. The maximal value depends on available sectors, so it should be <= length(indicator.digit_pins)

NumDisplay.write_number — Function

write_number(
    indicator::AbstractNumDisplay,
    number::Int,
    dp_position::Union{Int, Nothing} = nothing
)

Writes the decimal value to the display.

Arguments

indicator : object representing display device

number : decimal number. The maximal possible value depends on number of sectors and will be checked.

Example

# d is 4-digit (sector) display
write_number(d, 123)
# the result is _123

write_number(d, 1234)
# the result is 1234

write_number(d, 12345)
# throws an error

NumDisplay.write_number — Method

function write_number(
    indicator::DisplayBCD,
    digit_vector::AbstractArray{D},
    dp_position::Union(Int,Nothing) = nothing
) where D <: Union{UInt8, Nothing}

Writes several digits to the positions. The result of the execution is updating the whole number. If digit_vector is shorter than number of sectors the rest sectors will be empty.

Arguments

indicator : object representing display device

digitvector : vector of decimal values from 0 to 9 or nothing. The same meaning as digit in `writedigit()`. The first element of vector will be writte to the less significant sector, etc.

dp_position : position of dot in display

Example

# d is 4-digit (sector) display
write_number(d, [1,2])
# the result is __321

write_number(d, [1,2,3,4])
# the result is 4321

write_number(d, [1,2,3,4,5,6])
# the result is 4321